requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pycparser==2.22
PySocks==1.7.1
pytesseract==0.3.13
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
